+++ magda-minion-broken-link/src/wait.ts
@@ -1,0 +1,13 @@
+/**
+ * Wait for `waitMilliSeconds` before resolve the promise
+ */
+export default function wait(waitMilliSeconds: number): Promise<void> {
+    if (!waitMilliSeconds) {
+        return Promise.resolve();
+    }
+    return new Promise((resolve, reject) => {
+        setTimeout(() => {
+            resolve();
+        }, waitMilliSeconds);
+    });
+}